Move mongo update and query structs into query_selector module
